Determine which driver to use automatically
