+++ kchd/driver/__init__.py
@@ -3,6 +3,7 @@
 from typing import Set
 
 from kchd.hardware import KCHLED
+from kchd.types import NoKCHException
 
 from .driver import LEDDriver
 from .gpio import GPIODriver
@@ -10,7 +11,14 @@ from .gpio import GPIODriver
 
 def get_driver(leds: Set[KCHLED]) -> LEDDriver:
     """Get the driver to use."""
-    return GPIODriver(leds)
+    for driver_cls in [GPIODriver]:
+        driver = driver_cls(leds)
+        try:
+            driver.get_kch_info()
+            return driver
+        except NoKCHException:
+            pass
+    raise RuntimeError("No drivers were available.")
 
 
 __all__ = ["get_driver", "GPIODriver", "LEDDriver"]
