+++ kchd/driver/driver.py
@@ -3,6 +3,7 @@
 from typing import Dict, Protocol, Set
 
 from kchd.hardware import KCHLED
+from kchd.types import KCHInfo
 
 
 class LEDDriver(Protocol):
@@ -10,6 +11,14 @@ class LEDDriver(Protocol):
 
     def __init__(self, leds: Set[KCHLED]) -> None:
         """Initialise and set up the LEDs."""
+        ...
+
+    def get_kch_info(self) -> KCHInfo:
+        """
+        Get information about the KCH this driver operates.
+
+        :raises NoKCHException: There is no KCH available.
+        """
         ...
 
     def set_state(self, state: Dict[KCHLED, bool]) -> None:
